requirements.txt
python-telegram-bot==20.0
openai>=1.0
python-dotenv>=0.19
beautifulsoup4>=4.0
requests>=2.25
pytest>=7.0